doc: Use www.gitblit.com in link to full release notes
The release process generates a release on Github. The release notes
on Github link to the full release notes on the Gitblit website.
But since gitblit.com/ only has a simple redirect in the index.html to
the Github pages site, a direct link to the releasenotes on gitblit.com
will end up in the outdated pages on Jame's site.

Therefore use www.gitblit.com/ for the link, which we can control and
which leads to the Github pages directly.